Rename pause_aarch64.hpp to spin_wait_aarch64.hpp
